move anot her test to test old code. (mid_min_max)
